fix(scheduler): inline SchedulerDO in worker.ts to survive bundling
Builds at 07:28Z and 07:34Z deployed successfully per CF but all
routes returned HTTP 404 with x-preview-user-error: true — workerd
refused to start with "no such actor class; c = SchedulerDO". The
build-time warning "no such Durable Object class is exported from
the worker" turned out to be load-bearing, not benign.

Root cause: importing SchedulerDO from ./lib/scheduler/scheduler-do
and re-exporting from worker.ts didn't survive the OpenNext +
wrangler esbuild pipeline. The class was stripped from the deployed
bundle even though the export statement was present.

Fix: inline the class body directly in worker.ts so it lives at the
worker entry point. Removes the module boundary the bundler was
losing. Same pattern OpenNext's own DOs (DOQueueHandler etc.) use —
they live inside .open-next/worker.js, not behind a separate import.

Also point cloudflare-env.d.ts SCHEDULER typing at ./worker rather
than the deleted ./lib/scheduler/scheduler-do path.

Refs #768. Context: OpenNext-Cloudflare issue #502 for the broader
custom-DO bundling pattern.
