+++ cloudflare-env.d.ts
@@ -18,6 +18,6 @@ interface CloudflareEnv {
   X402_RELAY_URL?: string; // x402 relay URL for all payment settlement (default: https://x402-relay.aibtc.com)
   X402_RELAY?: import("./lib/inbox/relay-rpc").RelayRPC; // x402 sponsor relay RPC service binding (undefined in local dev)
   INBOX_RECONCILIATION_QUEUE?: Queue<import("./lib/inbox/reconciliation-queue").InboxReconciliationQueueMessage>;
-  SCHEDULER: DurableObjectNamespace<import("./lib/scheduler/scheduler-do").SchedulerDO>;
+  SCHEDULER: DurableObjectNamespace<import("./worker").SchedulerDO>;
   TENERO_API_KEY?: string; // Optional Tenero API key (x-api-key header); raises rate limits above the shared web-ui-ip tier
 }
